+++ runner/scheduler.py
@@ -19,9 +19,8 @@ from reporting.generate import generate_report
 from reporting.results import TestResult
 from runner.config import RunConfig
 from runner.master_log import MasterLog
-from utils.constants import EventType, ResetPolicy
+from utils.constants import EventType
 from utils.filling_task import filler_no_response_tokens_trivia
-from utils.text import token_len
 from utils.ui import colour_print
 from utils.files import make_runstats_path, make_master_log_path
 from runner.progress import ProgressDialog
@@ -157,8 +156,8 @@ class TestRunner:
             test_id=test_id, message=action.reply, timestamp=action.reply_ts,  is_question=action.is_question
         )
         self.agent_benchmark_duration += (action.reply_ts - action.sent_ts).total_seconds()
-        message_tokens = token_len(action.message)
-        reply_tokens = token_len(action.reply)
+        message_tokens = self.agent.token_len(action.message)
+        reply_tokens = self.agent.token_len(action.reply)
         self.agent_token_count += reply_tokens
         self.total_token_count += message_tokens + reply_tokens
         return message_tokens + reply_tokens
@@ -230,7 +229,7 @@ class TestRunner:
             num_tokens = self.wait_list[token_waiting_id]["tokens"]
             remaining_tokens = num_tokens - self.total_token_count
             while remaining_tokens > 0:
-                msg = filler_no_response_tokens_trivia(remaining_tokens, self.agent.max_message_size)
+                msg = filler_no_response_tokens_trivia(remaining_tokens, self.agent.max_message_size, self.agent)
                 tokens_spent = self.send_message("", SendMessageAction(msg, is_filling=True))
                 remaining_tokens -= tokens_spent
             if not self.is_waiting(token_waiting_id, remove=True):
@@ -282,7 +281,7 @@ class TestRunner:
                     self.set_to_wait(test, action, log_this=False)
                     self.reset_time()
 
-    def setup_iterator(self, test_group, reset_policy):
+    def setup_iterator(self, test_group):
         """Sets up the test dict and fast forwards any tests that are currently in progress"""
         tests = {t.unique_id: t for t in test_group}
 
@@ -291,11 +290,8 @@ class TestRunner:
             if isinstance(t, DynamicExample):
                 t.master_log = self.master_log
 
-        # Resuming a benchmark?
-        if reset_policy == ResetPolicy.SOFT:
-            self.fast_forward_tests(tests)
-        else:
-            pass  # `run_tests` will skip finished tests and start the remaining tests over.
+        # Fast forward tests to where they were
+        self.fast_forward_tests(tests)
 
         # Check if the last event in the log is after the current time, then travel to that time if it is.
         if len(self.master_log.log) > 0 and datetime.now() < self.master_log.log[-1].timestamp:
@@ -303,13 +299,13 @@ class TestRunner:
 
         # Add a reset event to the log if it has indeed been reset
         if len(self.master_log.log) > 0:
-            self.master_log.add_reset_event(self.agent.reset_policy, datetime.now())
+            self.master_log.add_reset_event(datetime.now())
 
         return tests
 
     def iter_tests(self, test_group: list[TestExample]) -> Iterator[TestExample]:
         # Set up the tests that are to be iterated through, fast forwarding where appropriate
-        tests = self.setup_iterator(test_group, self.agent.reset_policy)
+        tests = self.setup_iterator(test_group)
 
         assert len(tests) == len(test_group), f"There are tests with identical IDs: {[t.unique_id for t in test_group]}"
         while len(tests) > 0:
