+++ utils/text.py
@@ -1,10 +1,4 @@
-import tiktoken
 from rouge_score import rouge_scorer
-
-
-def token_len(text: str) -> int:
-    enc = tiktoken.get_encoding("cl100k_base")
-    return len(enc.encode(text))
 
 
 def rouge_l(reference: str, candidate: str) -> float:
